[NIE] New update scripts and lists

+++ tools/saber_compare.sh
@@ -21,36 +21,38 @@ if test "${test%%_*}" = "bump" ; then
    tolerance=1.e-4
 
    for file in `ls testdata/${test}/test_${mpi}-${omp}_*.nc` ; do
-      # Get suffix
-      tmp=${file#testdata/${test}/test_${mpi}-${omp}_}
-      suffix=${tmp%.nc}
+      if test ! -L ${file}; then
+         # Get suffix
+         tmp=${file#testdata/${test}/test_${mpi}-${omp}_}
+         suffix=${tmp%.nc}
 
-      if printf %s\\n "${suffix}" | grep -qF "distribution" ; then
-         # Remove distribution
-         rm -f ${file}
-      else
-         # Build reference file name, with a special case where the file is mpi-dependent
-         fileref=testref/${test}/test_1-1_${suffix}.nc
-         for special in "mom" "lct_cor" "nicas" "obs" "split" "vbal" ; do
-            if printf %s\\n "${suffix}" | grep -qF "${special}" ; then
-               fileref=testref/${test}/test_${mpi}-1_${suffix}.nc
-            fi
-         done
+         if printf %s\\n "${suffix}" | grep -qF "distribution" ; then
+            # Remove distribution
+            rm -f ${file}
+         else
+            # Build reference file name, with a special case where the file is mpi-dependent
+            fileref=testref/${test}/test_1-1_${suffix}.nc
+            for special in "mom" "lct_cor" "nicas" "normality" "obs" "split" "vbal" ; do
+               if printf %s\\n "${suffix}" | grep -qF "${special}" ; then
+                  fileref=testref/${test}/test_${mpi}-1_${suffix}.nc
+               fi
+            done
 
-         if [ -x "$(command -v nccmp)" ] ; then
-            # Compare files with NCCMP
-            echo "Command: nccmp -dfFmqS --threads=${nthreads} -T ${tolerance} ${file} ${fileref}"
-            nccmp -dfFmqS --threads=${nthreads} -T ${tolerance} ${file} ${fileref}
-            exit_code=$?
-            if test "${exit_code}" != "0" ; then
-               echo "\e[31mTest failed (nccmp) checking: "${file#testdata/}"\e[0m"
-               status=1
+            if [ -x "$(command -v nccmp)" ] ; then
+               # Compare files with NCCMP
+               echo "Command: nccmp -dfFmqS --threads=${nthreads} -T ${tolerance} ${file} ${fileref}"
+               nccmp -dfFmqS --threads=${nthreads} -T ${tolerance} ${file} ${fileref}
+               exit_code=$?
+               if test "${exit_code}" != "0" ; then
+                  echo "\e[31mTest failed (nccmp) checking: "${file#testdata/}"\e[0m"
+                  status=1
+                  exit ${status}
+               fi
+            else
+               echo "\e[31mCannot find command: nccmp\e[0m"
+               status=2
                exit ${status}
             fi
-         else
-            echo "\e[31mCannot find command: nccmp\e[0m"
-            status=2
-            exit ${status}
          fi
       fi
    done
